Decompile func0f01d990

+++ src/include/game/data/data_020df0.h
@@ -53,7 +53,7 @@ extern struct stagethinglist stagethinglist_22ec8;
 extern struct stagethinglist stagethinglist_22f1c;
 extern struct body g_Bodies[];
 extern u32 var8007dae4;
-extern u32 var8007db80;
+extern f32 var8007db80;
 extern f32 var8007db84;
 extern u32 var8007db88;
 extern u32 var8007db94;
